+++ tests/Unit/DateTimeTest.php
@@ -24,7 +24,7 @@ class DateTimeTest extends LaravelTableTestCase
         $table->column('name');
         $table->column('updated_at')->dateTimeFormat('d/m/Y');
         $table->render();
-        $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
+        $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString(Carbon::parse($user->updated_at)->format('d/m/Y'), $html);
     }
 
@@ -43,7 +43,7 @@ class DateTimeTest extends LaravelTableTestCase
         $table->column('name');
         $table->column('updated_at')->dateTimeFormat('d/m');
         $table->render();
-        $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
+        $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString(Carbon::parse($user->updated_at)->format('d/m'), $html);
     }
 
@@ -55,7 +55,7 @@ class DateTimeTest extends LaravelTableTestCase
         $table->column('name');
         $table->column('updated_at')->dateTimeFormat('H\h i\m\i\n');
         $table->render();
-        $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
+        $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString(Carbon::parse($user->updated_at)->format('H\h i\m\i\n'), $html);
     }
 
@@ -67,7 +67,7 @@ class DateTimeTest extends LaravelTableTestCase
         $table->column('name');
         $table->column('updated_at')->dateTimeFormat('d/m/Y H:i');
         $table->render();
-        $html = view('laravel-table::' . $table->tbodyComponentPath, compact('table'))->render();
+        $html = view('laravel-table::' . $table->tbodyTemplatePath, compact('table'))->render();
         $this->assertStringContainsString(Carbon::parse($user->updated_at)->format('d/m/Y H:i'), $html);
     }
 }