+++ tests/views/destroy-test.blade.php
@@ -1,0 +1,3 @@
+<form id="destroy-test"></form>
+
+