+++ tests/views/show-test.blade.php
@@ -1,0 +1,3 @@
+<form id="show-test"></form>
+
+