+++ tests/views/table-test.blade.php
@@ -1,7 +1,7 @@
 <div>
     <table id="table-test">
-        @include('laravel-table::' . $table->theadComponentPath)
-        @include('laravel-table::' . $table->tbodyComponentPath)
+        @include('laravel-table::' . $table->theadTemplatePath)
+        @include('laravel-table::' . $table->tbodyTemplatePath)
         @include('laravel-table::' . $table->tfootComponentPath)
     </table>
 </div>